Switching read byte counters to unsigned types to ensure we don't get int wrapping when counting the bytes read. Previously MAX_LINGER_BYTES (32768) was > INT_MAX.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/jk/trunk@1125651 13f79535-47bb-0310-9956-ffa450edef68

+++ native/common/jk_connect.c
@@ -712,8 +712,8 @@ int jk_shutdown_socket(jk_sock_t sd, jk_logger_t *l)
     char buf[64];
     char *sb = NULL;
     int rc = 0;
-    int rd = 0;
-    int rp = 0;
+    size_t rd = 0;
+    size_t rp = 0;
     int save_errno;
     int timeout = MS_TO_LINGER;
     time_t start = time(NULL);
